Move download into scripts as a callable function
